requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
altair==6.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.1.1
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cryptography==48.0.0
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.136.1
filetype==1.2.0
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.50
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.196.0
google-auth==2.51.0
google-auth-httplib2==0.4.0
google-genai==1.75.0
google-generativeai==0.8.6
googleapis-common-protos==1.75.0
greenlet==3.5.0
groq==0.37.1
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.13
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.17
langchain-classic==1.0.6
langchain-community==0.4.1
langchain-core==1.3.3
langchain-google-genai==4.2.2
langchain-groq==1.1.2
langchain-openai==1.2.1
langchain-protocol==0.0.15
langchain-text-splitters==1.1.2
langgraph==1.1.10
langgraph-checkpoint==4.0.3
langgraph-prebuilt==1.0.13
langgraph-sdk==0.3.14
langsmith==0.8.2
lxml==6.1.0
markdown-it-py==4.1.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.20.0
numpy==2.4.4
openai==2.35.1
orjson==3.11.9
ormsgpack==1.12.2
packaging==26.2
pandas==3.0.2
pillow==12.2.0
propcache==0.4.1
proto-plus==1.28.0
protobuf==5.29.6
pyarrow==24.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.4
pydantic-settings==2.14.0
pydantic_core==2.46.4
pydeck==0.9.2
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.27
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
rich==15.0.0
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.49
starlette==1.0.0
streamlit==1.57.0
tavily-python==0.7.24
tenacity==9.1.4
tiktoken==0.12.0
toml==0.10.2
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.46.0
watchdog==6.0.0
webencodings==0.5.1
websockets==16.0
xxhash==3.7.0
yarl==1.23.0
zstandard==0.25.0
